refactor(canonical): isolate lifecycle ownership
